+++ src/EventSubscriber/HomepageCookieLanguageRedirection.php
@@ -23,6 +23,20 @@ use Symfony\Component\HttpKernel\KernelEvents;
 class HomepageCookieLanguageRedirection implements EventSubscriberInterface {
 
   /**
+   * The Cookie Redirection must be triggered after the Browser redirection.
+   *
+   * The value here must be lower than
+   * {@HomepageBrowserLanguageRedirection::PRIORITY}.
+   * This needs to run after \Symfony\Component\HttpKernel\EventListener\RouterListener::onKernelRequest(),
+   * which has a priority of 32.
+   * This needs to run after \Drupal\home_redirect_lang\EventSubscriber\HomepageBrowserLanguageRedirection::redirectPreferredLanguage(),
+   * which has a priority of 31.
+   *
+   * @var int
+   */
+  private const PRIORITY = 30;
+
+  /**
    * Symfony\Component\HttpFoundation\RequestStack definition.
    *
    * @var \Symfony\Component\HttpFoundation\Request
@@ -63,9 +77,9 @@ class HomepageCookieLanguageRedirection implements EventSubscriberInterface {
   /**
    * {@inheritdoc}
    */
-  public static function getSubscribedEvents() {
+  public static function getSubscribedEvents(): array {
     return [
-      KernelEvents::REQUEST => ['redirectPreferredLanguage'],
+      KernelEvents::REQUEST => ['redirectPreferredLanguage', self::PRIORITY],
     ];
   }
 
@@ -75,18 +89,18 @@ class HomepageCookieLanguageRedirection implements EventSubscriberInterface {
    * @param \Symfony\Component\HttpKernel\Event\RequestEvent $event
    *   The event.
    */
-  public function redirectPreferredLanguage(RequestEvent $event) {
+  public function redirectPreferredLanguage(RequestEvent $event): void {
     if (!$this->pathMatcher->isFrontPage()) {
       return;
     }
 
-    // Whether or not preventing redirection when Referer Header is given.
+    // Whether preventing redirection when Referer Header is given.
     $referer_bypass_enabled = (bool) $this->configFactory->get('home_redirect_lang.cookie')->get('enable_referer_bypass');
 
     $current_language = $this->languageManager->getCurrentLanguage();
     $http_referer = $this->request->server->get('HTTP_REFERER');
     $current_host = $this->request->getHost();
-    $referer_host = parse_url($http_referer, \PHP_URL_HOST);
+    $referer_host = parse_url((string) $http_referer, \PHP_URL_HOST);
 
     // Ensure the REFERER is external to disable redirection.
     if ($referer_bypass_enabled && !empty($referer_host) && !empty($current_host) && $current_host !== $referer_host) {