+++ CHANGELOG.md
@@ -7,13 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - fix Issue #3320300: Avoid "Uncaught ReferenceError: Drupal is not defined" for anonymous users
+- fix parse_url(): passing null to parameter #1 () of type string is deprecated
+- fix Drupal 10 (Symfony 6) Kernel Event Priorities
 
 ### Removed
 - remove satackey/action-docker-layer-caching on Github Actions
+- drop support of drupal below 9.3.x
 
 ### Added
 - add coverage for Drupal 9.3, 9.4 & 9.5
-- drop support of drupal below 9.3.x
+- add official support of drupal 9.5 & 10.0
+
+### Changed
+- re-enable PHPUnit Symfony Deprecation notice
 
 ## [1.0.0-alpha1] - 2022-09-22
 ### Added